requirements.txt
prometheus_client
Flask
supervisor
gunicorn
pyyaml
fire